+++ estimagic/optimization/internal_criterion_template.py
@@ -6,7 +6,7 @@ import numpy as np
 from estimagic.differentiation.derivatives import first_derivative
 from estimagic.exceptions import get_traceback
 from estimagic.logging.database_utilities import append_row
-from estimagic.optimization.utilities import hash_array
+from estimagic.utilities import hash_array
 
 
 DERIVATIVE_ERROR_MESSAGE = (
