+++ estimagic/parameters/kernel_transformations.py
@@ -30,12 +30,12 @@ specific cases we refer to posts on math.stackexchange.com.
 """
 import numpy as np
 
-from estimagic.optimization.utilities import chol_params_to_lower_triangular_matrix
-from estimagic.optimization.utilities import cov_matrix_to_sdcorr_params
-from estimagic.optimization.utilities import cov_params_to_matrix
-from estimagic.optimization.utilities import dimension_to_number_of_triangular_elements
-from estimagic.optimization.utilities import robust_cholesky
-from estimagic.optimization.utilities import sdcorr_params_to_matrix
+from estimagic.utilities import chol_params_to_lower_triangular_matrix
+from estimagic.utilities import cov_matrix_to_sdcorr_params
+from estimagic.utilities import cov_params_to_matrix
+from estimagic.utilities import dimension_to_number_of_triangular_elements
+from estimagic.utilities import robust_cholesky
+from estimagic.utilities import sdcorr_params_to_matrix
 
 
 def covariance_to_internal(external_values, constr):
@@ -471,8 +471,8 @@ def _transformation_matrix(dim):
     Example:
     >>> import numpy as np
     >>> from numpy.testing import assert_array_almost_equal
-    >>> from estimagic.optimization.utilities import cov_matrix_to_sdcorr_params
-    >>> from estimagic.optimization.utilities import cov_to_sds_and_corr
+    >>> from estimagic.utilities import cov_matrix_to_sdcorr_params
+    >>> from estimagic.utilities import cov_to_sds_and_corr
     >>> cov = np.cov(np.random.randn(10, 4))
     >>> sds, corr = cov_to_sds_and_corr(cov)
     >>> corr[np.diag_indices(len(cov))] = sds
