+++ estimagic/utilities.py
